// generated/lvideo/v1/live_video.ts
export interface BroadcastConfig {
  enableHls: boolean;
  enableRecording: boolean;
}

export interface CreateRoomRequest {
  streamId: string;
  broadcasterId: string;
  /** tokens will be added in the headers */
  config: BroadcastConfig | undefined;
}

export interface CreateRoomResponse {
  success: boolean;
  message: string;
}

function createBaseBroadcastConfig(): BroadcastConfig {
  return { enableHls: false, enableRecording: false };
}

export const BroadcastConfig: MessageFns<BroadcastConfig> = {
  encode(message: BroadcastConfig, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.enableHls !== false) {
      writer.uint32(8).bool(message.enableHls);
    }
    if (message.enableRecording !== false) {
      writer.uint32(16).bool(message.enableRecording);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): BroadcastConfig {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseBroadcastConfig();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.enableHls = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.enableRecording = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): BroadcastConfig {
    return {
      enableHls: isSet(object.enableHls)
        ? globalThis.Boolean(object.enableHls)
        : isSet(object.enable_hls)
        ? globalThis.Boolean(object.enable_hls)
        : false,
      enableRecording: isSet(object.enableRecording)
        ? globalThis.Boolean(object.enableRecording)
        : isSet(object.enable_recording)
        ? globalThis.Boolean(object.enable_recording)
        : false,
    };
  },

  toJSON(message: BroadcastConfig): unknown {
    const obj: any = {};
    if (message.enableHls !== false) {
      obj.enableHls = message.enableHls;
    }
    if (message.enableRecording !== false) {
      obj.enableRecording = message.enableRecording;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<BroadcastConfig>, I>>(base?: I): BroadcastConfig {
    return BroadcastConfig.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<BroadcastConfig>, I>>(object: I): BroadcastConfig {
    const message = createBaseBroadcastConfig();
    message.enableHls = object.enableHls ?? false;
    message.enableRecording = object.enableRecording ?? false;
    return message;
  },
};

function createBaseCreateRoomRequest(): CreateRoomRequest {
  return { streamId: "", broadcasterId: "", config: undefined };
}

export const CreateRoomRequest: MessageFns<CreateRoomRequest> = {
  encode(message: CreateRoomRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.streamId !== "") {
      writer.uint32(10).string(message.streamId);
    }
    if (message.broadcasterId !== "") {
      writer.uint32(18).string(message.broadcasterId);
    }
    if (message.config !== undefined) {
      BroadcastConfig.encode(message.config, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateRoomRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateRoomRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.streamId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.broadcasterId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.config = BroadcastConfig.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateRoomRequest {
    return {
      streamId: isSet(object.streamId)
        ? globalThis.String(object.streamId)
        : isSet(object.stream_id)
        ? globalThis.String(object.stream_id)
        : "",
      broadcasterId: isSet(object.broadcasterId)
        ? globalThis.String(object.broadcasterId)
        : isSet(object.broadcaster_id)
        ? globalThis.String(object.broadcaster_id)
        : "",
      config: isSet(object.config) ? BroadcastConfig.fromJSON(object.config) : undefined,
    };
  },

  toJSON(message: CreateRoomRequest): unknown {
    const obj: any = {};
    if (message.streamId !== "") {
      obj.streamId = message.streamId;
    }
    if (message.broadcasterId !== "") {
      obj.broadcasterId = message.broadcasterId;
    }
    if (message.config !== undefined) {
      obj.config = BroadcastConfig.toJSON(message.config);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateRoomRequest>, I>>(base?: I): CreateRoomRequest {
    return CreateRoomRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateRoomRequest>, I>>(object: I): CreateRoomRequest {
    const message = createBaseCreateRoomRequest();
    message.streamId = object.streamId ?? "";
    message.broadcasterId = object.broadcasterId ?? "";
    message.config = (object.config !== undefined && object.config !== null)
      ? BroadcastConfig.fromPartial(object.config)
      : undefined;
    return message;
  },
};

function createBaseCreateRoomResponse(): CreateRoomResponse {
  return { success: false, message: "" };
}

export const CreateRoomResponse: MessageFns<CreateRoomResponse> = {
  encode(message: CreateRoomResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.success !== false) {
      writer.uint32(8).bool(message.success);
    }
    if (message.message !== "") {
      writer.uint32(18).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateRoomResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateRoomResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.success = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateRoomResponse {
    return {
      success: isSet(object.success) ? globalThis.Boolean(object.success) : false,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
    };
  },

  toJSON(message: CreateRoomResponse): unknown {
    const obj: any = {};
    if (message.success !== false) {
      obj.success = message.success;
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateRoomResponse>, I>>(base?: I): CreateRoomResponse {
    return CreateRoomResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateRoomResponse>, I>>(object: I): CreateRoomResponse {
    const message = createBaseCreateRoomResponse();
    message.success = object.success ?? false;
    message.message = object.message ?? "";
    return message;
  },
};

export type LiveVideoServiceService = typeof LiveVideoServiceService;
export const LiveVideoServiceService = {
  createRoom: {
    path: "/lvideo.v1.LiveVideoService/CreateRoom",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: CreateRoomRequest): Buffer => Buffer.from(CreateRoomRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): CreateRoomRequest => CreateRoomRequest.decode(value),
    responseSerialize: (value: CreateRoomResponse): Buffer => Buffer.from(CreateRoomResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): CreateRoomResponse => CreateRoomResponse.decode(value),
  },
} as const;

export interface LiveVideoServiceServer extends UntypedServiceImplementation {
  createRoom: handleUnaryCall<CreateRoomRequest, CreateRoomResponse>;
}

export interface LiveVideoServiceClient extends Client {
  createRoom(
    request: CreateRoomRequest,
    callback: (error: ServiceError | null, response: CreateRoomResponse) => void,
  ): ClientUnaryCall;
  createRoom(
    request: CreateRoomRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: CreateRoomResponse) => void,
  ): ClientUnaryCall;
  createRoom(
    request: CreateRoomRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: CreateRoomResponse) => void,
  ): ClientUnaryCall;
}

export const LiveVideoServiceClient = makeGenericClientConstructor(
  LiveVideoServiceService,
  "lvideo.v1.LiveVideoService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): LiveVideoServiceClient;
  service: typeof LiveVideoServiceService;
  serviceName: string;
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}